Move the optimize part of LCP into a new class
